backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
greenlet>=3.0.0
pydantic-settings>=2.2.0
pygithub>=2.3.0
gitpython>=3.1.0
chromadb>=0.5.0
litellm>=1.40.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
httpx>=0.27.0
